- insert some static_casts for numeric conversion and/or make doubles float. This gets rid of most warnings by VC 7.1, and is safer anyway.

+++ src/Shape_buildblock/EllipsoidalCylinder.cxx
@@ -151,7 +151,7 @@ float
 EllipsoidalCylinder:: 
 get_geometric_volume()const
  {
-   return (radius_x*radius_y*_PI*length);
+   return static_cast<float>(radius_x*radius_y*_PI*length);
  }
 
 Shape3D*
